# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for the ResearchHub research workspace.",
        "title": "ResearchHub API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "ResearchHub Support",
            "email": "support@researchhub.dev"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/annotations/{annotationID}": {
            "put": {
                "description": "Updates the note content, font style, and font size of an annotation.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "annotations"
                ],
                "summary": "Update annotation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Annotation UUID",
                        "name": "annotationID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated fields",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpdateAnnotationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Annotation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Permanently deletes an annotation.",
                "tags": [
                    "annotations"
                ],
                "summary": "Delete annotation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Annotation UUID",
                        "name": "annotationID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/documents/{documentID}/annotations": {
            "get": {
                "description": "Returns all annotations for a document ordered by page and creation time.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "annotations"
                ],
                "summary": "List annotations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Document UUID",
                        "name": "documentID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Annotation"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a highlight annotation with optional sticky note.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "annotations"
                ],
                "summary": "Create annotation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Document UUID",
                        "name": "documentID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Annotation data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateAnnotationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Annotation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/projects/{projectID}/chat": {
            "post": {
                "description": "Streams token-by-token AI responses over Server-Sent Events (SSE).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Stream AI chat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project UUID",
                        "name": "projectID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User message",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SSE stream",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.ChatRequest": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Summarize the key findings"
                }
            }
        },
        "api.CreateAnnotationRequest": {
            "type": "object",
            "properties": {
                "coordinates": {
                    "$ref": "#/definitions/models.Coordinates"
                },
                "font_size": {
                    "type": "string",
                    "example": "medium"
                },
                "font_style": {
                    "type": "string",
                    "example": "sans-serif"
                },
                "note_content": {
                    "type": "string"
                },
                "page_number": {
                    "type": "integer",
                    "example": 1
                },
                "selection_txt": {
                    "type": "string",
                    "example": "selected text"
                }
            }
        },
...
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid request body"
                }
            }
        },
...
        "api.UpdateAnnotationRequest": {
            "type": "object",
            "properties": {
                "font_size": {
                    "type": "string",
                    "example": "medium"
                },
                "font_style": {
                    "type": "string",
                    "example": "sans-serif"
                },
                "note_content": {
                    "type": "string"
                }
            }
        },
        "models.Annotation": {
            "type": "object",
            "properties": {
                "coordinates": {
                    "$ref": "#/definitions/models.Coordinates"
                },
                "created_at": {
                    "type": "string"
                },
                "deep_link": {
                    "description": "researchhub:// URI",
                    "type": "string"
                },
                "document_id": {
                    "type": "string"
                },
                "font_size": {
                    "type": "string"
                },
                "font_style": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "note_content": {
                    "type": "string"
                },
                "page_number": {
                    "type": "integer"
                },
                "selection_txt": {
                    "type": "string"
                }
            }
        },
        "models.Coordinates": {
            "type": "object",
            "properties": {
                "height": {
                    "type": "number"
                },
                "width": {
                    "type": "number"
                },
                "x": {
                    "type": "number"
                },
                "y": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}