# server/internal/db/generated/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package dbgen

import (
	"encoding/json"

	"github.com/jackc/pgx/v5/pgtype"
	"github.com/pgvector/pgvector-go"
)

type Annotation struct {
	ID           pgtype.UUID        `json:"id"`
	DocumentID   pgtype.UUID        `json:"document_id"`
	PageNumber   int32              `json:"page_number"`
	SelectionTxt string             `json:"selection_txt"`
	Coordinates  json.RawMessage    `json:"coordinates"`
	NoteContent  pgtype.Text        `json:"note_content"`
	FontStyle    string             `json:"font_style"`
	FontSize     string             `json:"font_size"`
	CreatedAt    pgtype.Timestamptz `json:"created_at"`
}

type DocChunk struct {
	ID          pgtype.UUID     `json:"id"`
	DocumentID  pgtype.UUID     `json:"document_id"`
	PageNumber  int32           `json:"page_number"`
	TextContent string          `json:"text_content"`
	Embedding   pgvector.Vector `json:"embedding"`
}

type Document struct {
	ID          pgtype.UUID        `json:"id"`
	ProjectID   pgtype.UUID        `json:"project_id"`
	FileName    string             `json:"file_name"`
	StoragePath string             `json:"storage_path"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
}

type Project struct {
	ID        pgtype.UUID        `json:"id"`
	Name      string             `json:"name"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}